calendar issue
a) Add the icon in the front of calendar week view and day view events.
b) Change the week view and day view events to a list, the events will
be list as several rows in each views. And all the day view events will
likes like a all_day view, which means the event will fill the plug-in
space entirely.
c) Add the click event for each view, month view click to week view and
week view can click to day view.
d) Add the create time for day view event.
e) Modify the size and position for the event tooltip.